updated index (see comment for further details on LOV)
Disclaimer: There is (was?) an outage of LinkedOpenVocabularies today,
website reported HTTP code 503. The LOV part was already
collected weeks before (in a another attempt to update the index) and
was reused to update the whole
index today. I don't think that there were any substantial
changes in LOV since then.